feat: move add virtual authenticator method to private method
