--- spec/support/virtual_authenticator_helper.rb
@@ -1,9 +1,0 @@
-module VirtualAuthenticatorHelper
-  def add_virtual_authenticator
-    options = ::Selenium::WebDriver::VirtualAuthenticatorOptions.new
-    options.user_verification = true
-    options.user_verified = true
-    options.resident_key = true
-    page.driver.browser.add_virtual_authenticator(options)
-  end
-end
